# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "tafa3olcard Backend API",
    "version": "1.0.0",
    "description": "API Documentation for the tafa3olcard backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/admin/clients": {
      "get": {
        "summary": "List admin clients",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Clients retrieved"
          }
        }
      },
      "post": {
        "summary": "Create admin client",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Client created"
          }
        }
      }
    },
    "/admin/clients/financial-movements": {
      "get": {
        "summary": "List all client financial movements",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Financial movements retrieved"
          }
        }
      }
    },
    "/admin/clients/special-prices": {
      "get": {
        "summary": "List all client product special prices",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Client special prices retrieved"
          }
        }
      }
    },
    "/admin/clients/special-prices/bulk-delete": {
      "patch": {
        "summary": "Soft delete selected client product special prices",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Client special prices soft deleted"
          }
        }
      }
    },
    "/admin/clients/special-prices/{specialPriceId}": {
      "patch": {
        "summary": "Update client product special price",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "specialPriceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client special price updated"
          }
        }
      }
    },
    "/admin/clients/{id}": {
      "get": {
        "summary": "Get admin client details",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client retrieved"
          }
        }
      },
      "patch": {
        "summary": "Update admin client",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client updated"
          }
        }
      },
      "delete": {
        "summary": "Soft delete admin client",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client soft deleted"
          }
        }
      }
    },
    "/admin/clients/{id}/movements": {
      "post": {
        "summary": "Create client financial movement",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Financial movement created"
          }
        }
      }
    },
    "/admin/clients/{id}/open-credit": {
      "patch": {
        "summary": "Update client open credit limit",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client open credit updated"
          }
        }
      }
    },
    "/admin/clients/{id}/levels": {
      "get": {
        "summary": "List client service levels",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client levels retrieved"
          }
        }
      }
    },
    "/admin/clients/{id}/levels/generate": {
      "post": {
        "summary": "Generate missing client service levels",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client levels generated"
          }
        }
      }
    },
    "/admin/clients/{id}/levels/{levelId}": {
      "patch": {
        "summary": "Update client service level",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "levelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client level updated"
          }
        }
      }
    },
    "/admin/clients/{id}/special-prices": {
      "get": {
        "summary": "List client product special prices",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client special prices retrieved"
          }
        }
      },
      "post": {
        "summary": "Create client product special price",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Client special price created"
          }
        }
      }
    },
    "/admin/clients/{id}/special-prices/bulk-delete": {
      "patch": {
        "summary": "Soft delete selected client product special prices",
        "tags": [
          "AdminClients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client special prices soft deleted"
          }
        }
      }
    },
...
  },
...
}